Create utility module eduid_idp.testing.
Avoids confusion with multiple sets of test users used in different
tests.
